resolve extent reprojection mismatch and remove premature geobox normalisation
